Centralise channel methods in runtime
