+++ oak/server/rust/oak_runtime/src/node/wasm.rs
@@ -305,7 +305,7 @@ impl WasmInterface {
             channels: channels?,
         };
 
-        writer.write(msg)?;
+        self.runtime.channel_write(writer, msg)?;
 
         Ok(())
     }
@@ -333,7 +333,7 @@ impl WasmInterface {
         self.validate_ptr(dest, dest_capacity)?;
         self.validate_ptr(handles_dest, handles_capcity * 8)?;
 
-        let msg = reader.try_read_message(dest_capacity as usize, handles_capcity as usize)?;
+        let msg = self.runtime.channel_try_read_message(reader, dest_capacity as usize, handles_capcity as usize)?;
 
         let (actual_length, actual_handle_count) = match &msg {
             None => (0, 0),
